requirements.txt
cffi==1.15.0
numpy==1.21.4
opencv-python==4.5.4.60
pycparser==2.21

Pillow~=8.4.0
matplotlib~=3.5.0